requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.5
django-crispy-forms==1.10.0
django-email-verification==0.1.0
django-heroku==0.3.1
django-qr-code==2.1.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
Pillow==8.1.0
psycopg2==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.5
qrcode==6.1
reportlab==3.5.59
segno==1.3.1
six==1.15.0
sqlparse==0.4.1
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
